Add documentation infrastructure (#48)
* docs: basic infrastructure

* docs: additional pages infrastructure

* docs: add new metadata clases to API docs

* fix: import ordering

* docs: move section headers into RST files

This avoids many docstring complaints raised by our linters.
Specifically, the first line of any docstring should be a short summary
(ending in '.') which should be followed by an empty line. This is
incompatible with wanting a Sphinx header at the top of the `__init__`
pages. So instead of changing our linter, we move the headers to the RST
files. This will also have the nice benefit of providing useful
summaries (which do not have to be identical to the page title) to be
included in the `autosummary` tables.

* docs: add missing intersphinx extension