docs: cleanup of the `sampler` module (#58)
* refactor: rename the file of the `POVMPubResult`

* docs: module-level descriptions for `povm_toolbox.sampler`

* docs: updates to `POVMPubResult`

* docs: updates to `POVMSamplerJob`

* docs: include POVMSamplerPubLike

The autodoc extension of Sphinx has some known limitations with respect
to resolving aliases like the one put in place here, but hopefully this
will "auto-fix" itself in the future.

* docs: updates to `POVMSamplerPub`

* docs: updates to `POVMSampler`

* docs: remove flaky output from doctest

* Linting

* fix: missing docstring errors

These methods inherit the docstrings. But `mypy` appears unable to pick
up on this. Overwriting the `D102` error for `ruff` is not possible,
because we do not always have it enabled and then ruff complains about
an unneeded `D102` ignore.

* fix: typos