Improvements to test coverage (#62)
* ci: minor coverage pragmas

Some lines cannot be covered in CI because we only test coverage with a
single Python version (even though the CI runs multiple of them, we just
don't recombine the coverage data). Thus, this commit labels some lines
as uncovered to exclude them as missing from the coverage report.

* test: the threshold argument of the marginal probabilities dual

* ci: some more coverage exclusions

* test: the QiskitRuntimeService integration

Closes #49

* ci: expose secrets via env variables

* ci: debug secret

* ci: expose environment secrets to tox

* test: setting a measurement_layout

* test: specifying specific parameter values using `loc`

* test: improve compose_circuit unittests

This improves the unittests of the `POVMImplementation.compose_circuit`
method by actually running more than a single unique measurement sample.

* test: POVMImplementation.compose_circuit with present TranspileLayout

* fix: remove redundant check

This `elif` is actually not needed because the `QuantumCircuit.layout`
can either be `None` or a `TranspileLayout`.