fix: properly skip action secret-dependent test (#123)
The Github action secrets are not available for branches running on
forks or for PRs coming from forks or non-collaborator contributors
(such as dependendabot or mergify). Therefore, we must skip tests
depending on such secrets.

This was attempted in the past, but the check only worked when the
environment variable was actually `None`, not when it was an empty
string. This commit addresses that.

(cherry picked from commit 8751189217b6706c5c06e52d4606c0e9651061c7)